requirements.txt
pandas
json
logging
pathlib
pyarrow
pickle
numpy
torch
torch_geometric
matplotlib
pyyaml